Move more decoding logic out of headers
